services/processor/requirements.txt
confluent-kafka>=2.3.0
psycopg2-binary>=2.9.9
nltk>=3.8.1
